AuthenticationService - Changed the name of the DEFAULT_JAAS_SPNEGO_CONFIG to DEFAULT_JAAS_LOGIN_CONFIG whose value is now jaasmine.login
JaasLoginFilter - correctly get the user's REMOTE_USER name and Principal from the logged in Subject if setRemoteUserOnLogin.

SPNegoFilter - changed the default value of the appName (this should most likely be removed as an initparam since JGSS has a specific requirement for this)

SimpleLogoutServlet - changed the default value of the appName

JaasmineHttpServletRequest - changed constructor to take the logged-in subject, getRemoteuser will pull the username from the subject, getUserPrincipal will pull the principal (the first one found) from the Subject.